Full Refactor
After some more work experience decided to fully refactor the way i start this app, ordered things in components and used a centralized pinia store for most of the stuff, which is still not ideal but times better. Added tabs to the side panel, selected object/s now can have their properties changed bidirectionally and dynamically in the properties tab, when nothing is selected it is used as canvas properties, for setting bg color, name for exports and stuff, some info and quick actions. Added copy/cut/paste functionality with buttons for it in the properties section(for selected object/s) and keyboard shortcuts, which there are also for delete, save, the new undo/redo functionality, the user can also zoom in and out with the buttons on top and with alt+scroll wheel, and panning with alt+mouse. There is json export/import functionality to save/load your work and png, higher res png and json options for image export. New canvas button that resets absolutely everything and asks nicely with a warning confirmation modal before doing so. Objects panel showing all of the objects on the canvas with an svg indication of it's shape(with some smart selection for some of them), name of the object(there is also a unnecessarily complex under the hood naming system for auto shape names and names of copies.. and copies of copies...), checkmark indicates if it is selected and for now let's you select only one from there(fabric.js is way too broken..) and below the list a count of all selected objects. A lot of new primitive shapes with their own mini indication buttons in the ShapeToolbar, and their own svg's for the objects tab.. with a slot open for what i'll have to add on the exam, it's 8am and i want to have at least around an hour to sleep, so that's it from me for now...